% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dispatch.R
\name{Dispatch}
\alias{Dispatch}
\title{Dispatch R6 Class}
\description{
Send messages to jamovi  tables from anywhere in the code.
Warnings can be sent to tables or arrays. For tables in Array, they do not need to exist at
the moment the message is broadcast, because the message is stored in the parent object
and then passed to the table at run time.

Warnings can be transient (get removed after .init()) when \code{init=TRUE} is passed.

Errors must be sent to existing (already defined) objects. Errors are passed directly
to the jamovi object. If option \code{final=TRUE}, a \code{stop()} is issued
@export
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{warnings}}{Set a warning message to a jamovi table
It accepts a list with fields:
\itemize{
\item \code{topic} the nickname of a table in the format:
\itemize{
\item \code{tablename} a name of the table as set in the \code{r.yaml} jamovi file
\item \code{groupname_tablename}  the group the table belongs to and the table name
\item \code{arrayname} the array of the tables target of the message
}
\item \code{message} the message to broadcast
\item \code{init} (optional) \code{init=TRUE} makes the message transient (disappears after init)
\item \code{head} (optional)  it can be \code{info}, \code{warning}, and \code{error}. In jamovi, if \code{tablename}
is a table, does not have an effect. If \code{tablename} is a \code{Htlm} obj, it gives
the message a different format and color. In R, \code{warning} is issued as an R
warning, the others are added as notes to the tables.
}}

\item{\code{errors}}{set an error message to a jamovi table
It accepts a list with fields:
\itemize{
\item \code{topic} the nickname of a table in the format:
\itemize{
\item \code{tablename} a name of the table as set in the \code{r.yaml} jamovi file
\item \code{groupname_tablename}  the group the table belongs to and the table name
\item \code{arrayname} the array of the tables target of the message
}
\item \code{message} the message to broadcast
\item \code{final} (optional) \code{final=TRUE} makes module stop computation immediately
}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-Dispatch-new}{\code{Dispatch$new()}}
\item \href{#method-Dispatch-translate}{\code{Dispatch$translate()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dispatch-new"></a>}}
\if{latex}{\out{\hypertarget{method-Dispatch-new}{}}}
\subsection{Method \code{new()}}{
Initialize the Dispatch object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dispatch$new(results)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{results}}{A jamovi results object  \code{\link[jmvcore]{jmvcore::Results}}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-Dispatch-translate"></a>}}
\if{latex}{\out{\hypertarget{method-Dispatch-translate}{}}}
\subsection{Method \code{translate()}}{
Translate a message
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Dispatch$translate(msgobj)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{msgobj}}{a list containing the field \code{message}. The function looks in the environment
for a global list named \code{TRANS_WARNS}. The list must contain lists with the field \code{original} and
one of the following: \code{new}: a new message to replace \code{original};
\code{sub}: a string to substitute the \code{original} string. \code{append}: a string to append to \code{original}.
\code{prepend}: a string to prepend to \code{original}. The field are evaluated in this order.
To remove a warning or error, simple set \code{new=NULL}}
}
\if{html}{\out{</div>}}
}
}
}
