% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTabs.R
\name{SmartArray}
\alias{SmartArray}
\title{SmartArray R6 Class}
\description{
A programmer interface to init and polutate \code{\link[jmvcore]{jmvcore::Array}}
}
\section{Super class}{
\code{\link[jmvScaffold:SmartTable]{jmvScaffold::SmartTable}} -> \code{SmartArray}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{childrenObjs}}{a list of children \code{jmvScaffold::SmartTable}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmartArray-new}{\code{SmartArray$new()}}
\item \href{#method-SmartArray-initTable}{\code{SmartArray$initTable()}}
\item \href{#method-SmartArray-runTable}{\code{SmartArray$runTable()}}
\item \href{#method-SmartArray-retrieveNotes}{\code{SmartArray$retrieveNotes()}}
\item \href{#method-SmartArray-setNotes}{\code{SmartArray$setNotes()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvScaffold" data-topic="SmartTable" data-id="ci"><a href='../../jmvScaffold/html/SmartTable.html#method-SmartTable-ci'><code>jmvScaffold::SmartTable$ci()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvScaffold" data-topic="SmartTable" data-id="setColumnTitle"><a href='../../jmvScaffold/html/SmartTable.html#method-SmartTable-setColumnTitle'><code>jmvScaffold::SmartTable$setColumnTitle()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-new}{}}}
\subsection{Method \code{new()}}{
Initialize the SmartArray object, creating the children tables
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$new(array, estimator = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{array}}{A jamovi array of class \code{\link[jmvcore]{jmvcore::Array}}}

\item{\code{estimator}}{optional \link[R6:R6Class]{R6}. If set, SmartArray inquires
the R6 class for init and run functions to initialize and populate the jamovi tables.
the R6 class should provide functions using the convention \code{init_[nickname]} and \verb{run_[nickname] where the }nickname` is the unique identifier given to the jamovi table. The unique identifiers can be
\itemize{
\item \strong{arrayname}:            the name of the \code{\link[jmvcore]{jmvcore::Array}}
\item \strong{groupname_arrayname}:   the name of the \code{\link[jmvcore]{jmvcore::Group}} containing the \code{\link[jmvcore]{jmvcore::Array}}  and the Array name
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-initTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-initTable}{}}}
\subsection{Method \code{initTable()}}{
initializes all children tables from \code{runSource} or \code{estimator::run_[tablename]}, and
takes care of all aesthetic adjustments required
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$initTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-runTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-runTable}{}}}
\subsection{Method \code{runTable()}}{
fills all children tables from \code{runSource} or \code{estimator::run_[tablename]}, and
takes care of all aesthetic adjustments required
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$runTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-retrieveNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-retrieveNotes}{}}}
\subsection{Method \code{retrieveNotes()}}{
used internally, but must be kept public
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$retrieveNotes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-setNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-setNotes}{}}}
\subsection{Method \code{setNotes()}}{
set a note to all children tables. Not necessary if a \code{Dispatch} object is used instead
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$setNotes(note)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{note}}{a string: the message to set}
}
\if{html}{\out{</div>}}
}
}
}
