% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTabs.R
\name{SmartTable}
\alias{SmartTable}
\title{SmartTable R6 Class}
\description{
A programmer interface to init and polutate \code{\link[jmvcore]{jmvcore::Table}}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{name}}{the name of the \code{\link[jmvcore]{jmvcore::Table}}}

\item{\code{table}}{\code{\link[jmvcore]{jmvcore::Table}}}

\item{\code{topics}}{the different topics}

\item{\code{nickname}}{the unique identifier of the table used by other classes
to dispatch messages to the table or to interact with it}

\item{\code{expandOnInit}}{if TRUE, the jamovi table is expanded columnwise
if the data.frame passed in the init are more than the ones defined
in .r.yaml file. If FALSE (default) additional columns passed on init
are ignored}

\item{\code{expandOnRun}}{if TRUE, the jamovi table is expanded columnwise
if the data.frame passed in to the run function are more than the ones defined
in .r.yaml file.  If FALSE (default) additional columns passed on run
are ignored.}

\item{\code{expandSuperTitle}}{a character string. If not NULL, the expanded columns are
supertitled with the string}

\item{\code{expandFrom}}{from which column should the table be expanded}

\item{\code{activated}}{SmartTable are populated (run) if their source
is not NULL and they are visible. jamovi tables with \code{visible: false}
in the .r.yaml file can be make visible and activated with this option TRUE}

\item{\code{key}}{This should be explained in a vignette}

\item{\code{keys_sep}}{This should be explained in a vignette}

\item{\code{keys_raise}}{This should be explained in a vignette}

\item{\code{spaceBy}}{a vector of columns names, \code{\link[jmvcore]{jmvcore::Table}} columns can add extra space
between rows to enhance readability.  Passing one or more column names put an extra padding
when a new value is fond in the column.}

\item{\code{spaceAt}}{a vector of rows indices, \code{\link[jmvcore]{jmvcore::Table}} columns can add extra space
between rows to enhance readability.  An extra space will be added after the rows with the provided
indices.}

\item{\code{spaceMethod}}{I do not remeber what this does, but it was cool}

\item{\code{indent}}{a vector of row indices. Indent (shift a bit on the right) the value of
the first column of each row index provided}

\item{\code{combineBelow}}{a vector of column names or column indices. \code{\link[jmvcore]{jmvcore::Table}} columns can hide repeated consecutive
values in a column by setting \verb{combineBelow:} column property in r.yaml file. This option allows
to set this properties programmatically on execution. The special key \verb{new!} means that all new
columns created at init or run time should be combineBelow.}

\item{\code{ci_info}}{this requires a vignette}

\item{\code{columnTitles}}{a named list of the form list(name=label). Change the title of
a jamovi table column named \code{name} to \code{label}.}
}
\if{html}{\out{</div>}}
}
\section{Active bindings}{
\if{html}{\out{<div class="r6-active-bindings">}}
\describe{
\item{\code{initSource}}{If \code{estimator} is not passed to initialize()
set the source of the init() data used to initialize the table.
if can be:
\itemize{
\item a data.frame
\item a list of named lists
\item a function returning a data.frame or a list of named lists
}}

\item{\code{runSource}}{If \code{estimator} is not passed to initialize()
set the source of the run() data used to initialize the table.
if can be:
\itemize{
\item a data.frame
\item a list of named lists
\item a function returning a data.frame or a list of named lists
}}

\item{\code{activateOnData}}{If TRUE, check if the table got some data. For tables with
\code{visible: false} is the r.yaml file, makes them visible
if data are passed to it                             #'}

\item{\code{title}}{a string. Set the title of the jamovi table}

\item{\code{superTitle}}{a named list list(name=value). Set the superTitle
of \code{name} column to \code{value}}

\item{\code{setColumnVisible}}{a list of column names to set visible}

\item{\code{hideOn}}{a named list list(name=value)
for columns in names(list(name=value))
hide the column if all its cells are equal to
value. Value it is typically \code{Inf}, \code{NA} or \code{NaN}}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmartTable-new}{\code{SmartTable$new()}}
\item \href{#method-SmartTable-initTable}{\code{SmartTable$initTable()}}
\item \href{#method-SmartTable-runTable}{\code{SmartTable$runTable()}}
\item \href{#method-SmartTable-ci}{\code{SmartTable$ci()}}
\item \href{#method-SmartTable-setTopics}{\code{SmartTable$setTopics()}}
\item \href{#method-SmartTable-retrieveNotes}{\code{SmartTable$retrieveNotes()}}
\item \href{#method-SmartTable-setNotes}{\code{SmartTable$setNotes()}}
\item \href{#method-SmartTable-setColumnTitle}{\code{SmartTable$setColumnTitle()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-new}{}}}
\subsection{Method \code{new()}}{
Initialize the SmartTable object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$new(table, estimator = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{table}}{A jamovi table of class \code{\link[jmvcore]{jmvcore::Table}}}

\item{\code{estimator}}{optional \link[R6:R6Class]{R6}. If set, SmartTable inquires
the R6 class for init and run functions to initialize and populate the jamovi Table.
the R6 class should provide functions using the convention \code{init_[nickname]} where
the \code{nickname} is the unique identifier given to the jamovi table. The unique identifiers can be
\itemize{
\item \strong{tablename}:            the name of the \code{\link[jmvcore]{jmvcore::Table}} in jamovi .r.yaml file
\item \strong{groupname_tablename}:  the name of the \code{\link[jmvcore]{jmvcore::Group}} parenting the table and the table name
\item \strong{arrayname_tablename}:  the name of the \code{\link[jmvcore]{jmvcore::Array}}  parenting the table and the table name
}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-initTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-initTable}{}}}
\subsection{Method \code{initTable()}}{
Initialize the table. Check if data are passed, check if
the table need to be expanded columnwise, if rows should be added and if
messages are broadcast to go into errors or warnings. Show the table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$initTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-runTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-runTable}{}}}
\subsection{Method \code{runTable()}}{
Populates the table. Check if data are passed, check if
the table need to be expanded columnwise, if rows should be added and if
messages are broadcast to go into errors or warnings. Show the table.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$runTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-ci"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-ci}{}}}
\subsection{Method \code{ci()}}{
insert a superTitle over the confidence interval columns
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$ci(
  aroot,
  width = 95,
  label = NULL,
  format = "{}\% Confidence Intervals"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{aroot}}{the root of the name of the confidence bounds columns. If one names the
confidence bounds columns as \code{some.ci.lower} and \code{some.ci.upper}, and sets
\code{aroot} as \code{some}, the two columns will be superTitled with \verb{width \% Confidence Intervals}.}

\item{\code{width}}{the width of the CI to appear in the superTitle}

\item{\code{label}}{add labels to the superTitle}

\item{\code{format}}{default is \code{"{}\% Confidence Intervals"}, can be changed to pass different
superTitle formats}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-setTopics"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-setTopics}{}}}
\subsection{Method \code{setTopics()}}{
this should probabily fo to private
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$setTopics()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-retrieveNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-retrieveNotes}{}}}
\subsection{Method \code{retrieveNotes()}}{
this should probabily fo to private
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$retrieveNotes(dispatcher = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{the}}{Dispatcher class object to dispatch messages}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-setNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-setNotes}{}}}
\subsection{Method \code{setNotes()}}{
set a note (footnote) to be displayed under  the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$setNotes(dispatcher = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartTable-setColumnTitle"></a>}}
\if{latex}{\out{\hypertarget{method-SmartTable-setColumnTitle}{}}}
\subsection{Method \code{setColumnTitle()}}{
Change the title of the jamovi table programmatically
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartTable$setColumnTitle(name, title)}\if{html}{\out{</div>}}
}

}
}
