% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SmartTabs.R
\name{SmartArray}
\alias{SmartArray}
\title{SmartArray R6 Class}
\description{
A programmer interface to init and populutate \code{\link[jmvcore]{jmvcore::Array}}.

It works exactly like \link[jmvScaffold]{SmartTable} class, but it passes
all settings to the children tables
}
\section{Super class}{
\code{\link[jmvScaffold:SmartTable]{jmvScaffold::SmartTable}} -> \code{SmartArray}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SmartArray-new}{\code{SmartArray$new()}}
\item \href{#method-SmartArray-initTable}{\code{SmartArray$initTable()}}
\item \href{#method-SmartArray-runTable}{\code{SmartArray$runTable()}}
\item \href{#method-SmartArray-retrieveNotes}{\code{SmartArray$retrieveNotes()}}
\item \href{#method-SmartArray-setNotes}{\code{SmartArray$setNotes()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="jmvScaffold" data-topic="SmartTable" data-id="ci"><a href='../../jmvScaffold/html/SmartTable.html#method-SmartTable-ci'><code>jmvScaffold::SmartTable$ci()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvScaffold" data-topic="SmartTable" data-id="setColumnTitle"><a href='../../jmvScaffold/html/SmartTable.html#method-SmartTable-setColumnTitle'><code>jmvScaffold::SmartTable$setColumnTitle()</code></a></span></li>
<li><span class="pkg-link" data-pkg="jmvScaffold" data-topic="SmartTable" data-id="setTopics"><a href='../../jmvScaffold/html/SmartTable.html#method-SmartTable-setTopics'><code>jmvScaffold::SmartTable$setTopics()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-new"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$new(tables, estimator = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-initTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-initTable}{}}}
\subsection{Method \code{initTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$initTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-runTable"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-runTable}{}}}
\subsection{Method \code{runTable()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$runTable()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-retrieveNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-retrieveNotes}{}}}
\subsection{Method \code{retrieveNotes()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$retrieveNotes()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SmartArray-setNotes"></a>}}
\if{latex}{\out{\hypertarget{method-SmartArray-setNotes}{}}}
\subsection{Method \code{setNotes()}}{
set a note (footnote) to be displayed under the table
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SmartArray$setNotes(dispatcher = NULL)}\if{html}{\out{</div>}}
}

}
}
